Refactored Modix.Handlers.InvitePurgeHandler into Modix.Services.Moderation.ModerationInvitePurgingBehavior. This brings the feature under the umbrella of the ModerationActions table (I.E. deleted messages are logged the same way infractions are). Also changed authorization for the feature over to the claims-based authorization system. This effectively eliminates the need for thie feature to have any configuration stored through the BehaviourConfiguration system.
Fixed a bug where ModerationLoggingBehavior was not posting the ID of infractions to the log channel. Instead, it was posting the ID of the moderation action item (I.E. the ID of the log item itself, which is basically useless).

Refactored BehaviorBase.SelfExecuteOnScopedServiceAsync (again) to the much more general SelfExecuteRequest, with more potential overloads, and added top-level exception logging.