Remove unnecessary dependency in Mod Invite Purging
This dependency was refactored to be unused and shouldn't work as expected because Mod Invite Purging Behavior is behavior lifetime while the used service is request scope lifetime. It seems on some machines GetServices is returning something unexpected and injecting that. It looks like this is a bug in the DI library.