Add more logging, improve exception serialization
This adds logging for DiscordAdapter, as well as a bit more logging in ModerationInvitePurgeHandler.cs, which should hopefully help us track down issues in the invite purger.

I also added additional logging to Program.cs for various config options, and improved exception serialization for the discord log sink by switching from a blacklist to a whitelist (may need to be expanded in the future).