Workaround for vue-good-table regression
The component tries to apply additional sorting logic to number and text columns regardless of the passed sorting function, causing errors and a failure to sort on some columns - setting the type to date bypasses this (we use our own sorting function for those columns anyway).

Note to self: extract common behavior of these views.