Update Invite Links
This pr is one I said I'd make about last month or so. Since we got our discord.gg link, I figured we would use it here. This should be a clean PR - no changes are being made that affects the visible functions and everything should continue to operate as expected.