Refactored all functionality for DataDog stats pertaining to "user metrics" within member guilds into a `UserMetrics` namespace, and a `UserMetricsBehavior` class.
The one exception is the "infractions" stat, for which there is no appropriate Notification available for the behavior to listen for. This can be moved into `UserMetricsBehavior` when the infractions system is refactored to operate under the umbrella of Messaging.

All other usages of DataDog are "performance" metrics, rather than "user" metrics.

Implemented logging and testing for all UserMetrics functionality.