Refactored the Hosting Behavior system a bit. (#798)
I'm in the middle of writing a new behavior, which forced me to add some new functionality to the hosting system (starting and stopping notifications), and that forced me to look at the behavior system again. In retrospect, it's a bit convoluted, so I've simplified it a bit.

Happy to discuss.