Implemented various refactorings and enhancements to the Messaging system.
Removed the `INotification` interface, which serves literally no purpose.

Removed support for `IRequest` and `IRequestHandler`, as it's not currently used anywhere. This was originally intended to provide feature-parity with MediatR, but as it hasn't been used at all in over a year (and there's question of whether we'll use MediatR itself for this functionality) it makes no sense it keep it in here.

Added support for cancellation within `MessageDispatcher.Dispatch()`, to help prevent deadlocks from killing the whole bot. The `Dispatch()` method now supports an optional `timeout` paramter, and pulls a default value from configuration, when none is given.

Enhanced logging within `MessageDispatcher` and `MessagePublisher`.

Added support for self-logging notifications, through an `ILogScopeProvider` interface. Now, notifications can be setup to emit their state to an `ILogger` by creating a logger scope to inject this state into all log messages generated during publishing of the notification. Both `MessageDispatcher` and `MessagePublisher` will create said scope for any notification that implements `ILogScopeProvider`.